+++ src/main/java/com/WebSite/repo/FoodRepository.java
@@ -3,6 +3,7 @@ package com.WebSite.repo;
 import com.WebSite.models.Food;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FoodRepository extends CrudRepository<Food,Long>   {
+public interface FoodRepository extends CrudRepository<Food, Integer>
+{
 }
