+++ src/main/java/com/WebSite/repo/ImageRepository.java
@@ -3,5 +3,6 @@ package com.WebSite.repo;
 import com.WebSite.models.Image;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ImageRepository extends CrudRepository<Image,Long> {
+public interface ImageRepository extends CrudRepository<Image, Integer>
+{
 }
